Break up test_utils.py and refactor tests
Remove the use of reset_mock() and create separate tests.
